requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
click==7.0
django==2.1.7
djangorestframework==3.9.1
entrypoints==0.3
isort==4.3.10
mccabe==0.6.1
numpy==1.16.2
pandas==0.24.1
python-dateutil==2.8.0
pytz==2018.9
six==1.12.0
